Crate amethyst_audio

Loading and playing of audio files.

Modules

output
Provides structures and functions used to get audio outputs.

Structs

AudioBundle
Audio bundle

AudioEmitter
An audio source, add this component to anything that emits sound.

AudioListener
An audio listener, add this component to the local player character.

AudioSink
This structure provides a way to programmatically pick and play music.

AudioSystem
Syncs 3D transform data with the audio engine to provide 3D audio.

DecoderError
An error occurred while decoding the source.

DjSystem
Calls a closure if the AudioSink is empty.

FlacFormat
Loads audio from Flac files.

OggFormat
Loads audio from Ogg Vorbis files

Source
A loaded audio file

WavFormat
Loads audio from wav files.

Type Definitions

SourceHandle
A handle to a source asset.