Module amethyst_audio::output

Provides structures and functions used to get audio outputs.

Structs

Output
A speaker(s) through which audio can be played.

OutputIterator
An iterator over outputs

Functions

default_output
Get the default output, returns none if no outputs are available.

init_output
Initialize default output

outputs
Get a list of outputs available to the system.