Struct amethyst_audio::output::Output

pub struct Output { /* fields omitted */ }

A speaker(s) through which audio can be played.

By convention, the default output is stored as a resouce in the World.

Methods

impl Output

pub fn name(&self) -> String

Gets the name of the output

pub fn try_play_once(
&self,
source: &Source,
volume: f32
) -> Result<(), DecoderError>

Play a sound once. A volume of 1.0 is unchanged, while 0.0 is silent.

This will return an Error if the loaded audio file in source could not be decoded.

pub fn play_once(&self, source: &Source, volume: f32)

Play a sound once. A volume of 1.0 is unchanged, while 0.0 is silent.

This may silently fail, in order to get error information use try_play_once.

pub fn play_n_times(&self, source: &Source, volume: f32, n: u16)

Play a sound n times. A volume of 1.0 is unchanged, while 0.0 is silent.

This may silently fail, in order to get error information use try_play_n_times.

pub fn try_play_n_times(
&self,
source: &Source,
volume: f32,
n: u16
) -> Result<(), DecoderError>

Play a sound n times. A volume of 1.0 is unchanged, while 0.0 is silent.

This will return an Error if the loaded audio file in source could not be decoded.