Struct amethyst_audio::AudioListener

pub struct AudioListener {
    pub output: Output,
    pub left_ear: Point3<f32>,
    pub right_ear: Point3<f32>,
}

An audio listener, add this component to the local player character.

Fields
output: Output

Output used by this listener to emit sounds to

left_ear: Point3<f32>

Position of the left ear relative to the global transform on this entity.

right_ear: Point3<f32>

Position of the right ear relative to the global transform on this entity.