Struct amethyst_audio::AudioSink

pub struct AudioSink { /* fields omitted */ }

This structure provides a way to programmatically pick and play music.

Methods

impl AudioSink

pub fn new(output: &Output) -> AudioSink

Creates a new AudioSink using the given audio output.

pub fn append(&self, source: &Source) -> Result<(), DecoderError>

Adds a source to the sink's queue of music to play.

pub fn empty(&self) -> bool

Returns true if the sink has no more music to play.

pub fn volume(&self) -> f32

Retrieves the volume of the sink, between 0.0 and 1.0;

pub fn set_volume(&mut self, volume: f32)

Sets the volume of the sink.

pub fn play(&self)

Resumes playback of a paused sink. Has no effect if this sink was never paused.

pub fn pause(&self)

Pauses playback, this can be resumed with AudioSink::play

pub fn is_paused(&self) -> bool

Returns true if the sink is currently paused.

pub fn stop(&self)

Empties the sink's queue of all music.