Struct amethyst_audio::DjSystem

pub struct DjSystem<F, R> { /* fields omitted */ }

Calls a closure if the AudioSink is empty.

Methods

impl<F, R> DjSystem<F, R>

pub fn new(f: F) -> Self

Creates a new DjSystem with the music picker being f. The closure takes a parameter, which needs to be a reference to a resource type, e.g. &MusicLibrary. This resource will be fetched by the system and passed to the picker.